Refactor SubscriberListQuery
Split the old code into multiple files each with
a specific function. Move the query code out of
notification worker into SubscriberListQuery.
